Allow defining to_create per FactoryGirl.define block
This introduces state to the FactoryGirl::Syntax::Methods::DSL for
storing a defined to_create block.

When running a factory, we first look to see if there's a custom
to_create on the factory definition itself - if there is, we use it. If
there's no custom to_create, precedence is granted to the global
to_create, falling back to the default to_create when necessary.